get_time

+++ philo/tools.c
@@ -17,7 +17,7 @@ long long	get_time_ms(void)
 	struct timeval	tv;
 
 	gettimeofday(&tv, NULL);
-	return ((tv.tv_sec * 1000) + (tv.tv_usec / 1000));
+	return ((long long)tv.tv_sec * 1000 + tv.tv_usec / 1000);
 }
 
 void	precise_sleep(long long duration, t_data *data)
